+++ .env.example
@@ -1,2 +1,1 @@
 REACT_APP_BACKEND_URL=
-TEST_JWT=
