+++ src/components/Adminpage/index.jsx
@@ -1,106 +1,88 @@
 import { React, useState, useEffect } from 'react';
-import {
-  Container, Accordion, Button, Modal
-} from 'react-bootstrap';
-import styles from './admin.module.css';
-import { UpdateForm } from './UpdateForm';
-import { Login } from './Login';
+import { Container, Form, Button } from 'react-bootstrap';
+import axios from 'axios';
+import { Dashboard } from './Dashboard';
 
 export function Adminpage() {
+  const [isSignedIn, setIsSignedIn] = useState(false);
+  const [rollno, setRollNo] = useState();
+  const [pass, setPass] = useState();
+  const [isWrong, setIsWrong] = useState(false);
+  const jwtToken = localStorage.getItem('token');
   const backendUrl = process.env.REACT_APP_BACKEND_URL;
-  const [isLoading, setIsLoading] = useState(true);
-  const [clusters, setClusters] = useState([]);
-  const [show, setShow] = useState(false);
-  const [eventData, setEventData] = useState({});
-
-  const handleClose = () => setShow(false);
-  const handleShow = (event) => {
-    const tmpEventData = {
-      name: event.name,
-      is_reg_completed: event.is_reg_completed,
-      is_event_completed: event.is_event_completed,
-      rules: event.rules,
-      description: event.description,
-      form_link: event.form_link,
-      event_link: event.event_link,
-      image_link: event.image_link,
-      start_time: event.start_time,
-      end_time: event.end_time,
-      points: event.points
-    };
-    setEventData(tmpEventData);
-    setShow(true);
-  };
-
-  const fetchData = () => {
-    fetch(`${backendUrl}events/`)
-      .then((res) => res.json())
-      .then((data) => {
-        setClusters(data);
-        setIsLoading(false);
-      });
-  };
+  if (!isSignedIn && jwtToken != null) setIsSignedIn(true);
   useEffect(() => {
-    fetchData();
-  }, []);
-
+  }, [isSignedIn]);
+  async function loginUser(credentials) {
+    return axios
+      .post(`${backendUrl}admin/`, JSON.stringify(credentials), {
+        headers: {
+          'Content-Type': 'application/json'
+        }
+      })
+      .then((response) => {
+        if (response.data.isAuthorized === 'True') {
+          localStorage.setItem('token', response.data.jwt_token);
+          setIsSignedIn(false);
+          setIsWrong(false);
+        } else {
+          setIsWrong(true);
+        }
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  }
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    await loginUser({
+      roll_number: rollno,
+      password: pass
+    });
+  };
+  const showIsWrongText = () => {
+    if (isWrong) {
+      return (
+        <h6 className="mb-4" style={{ color: '#FF0000' }}>
+          Wrong Password
+        </h6>
+      );
+    }
+    return null;
+  };
   return (
     <Container className="mt-5 mb-3">
-      <div className="m-auto w-50">
-        <h1 className="mb-4" style={{ color: '#6AE1DD' }}>
-          Admin Page
-        </h1>
-        <Modal className={styles.form} show={show} onHide={handleClose}>
-          <Modal.Header closeButton>
-            <Modal.Title>Update Event</Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            <UpdateForm eventData={eventData} />
-          </Modal.Body>
-        </Modal>
-        <Login />
-        {isLoading ? (
-          <div>Loading</div>
-        ) : (
-          <div>
-            {clusters
-              && clusters.map((cluster) => (
-                <Accordion key={cluster.cluster}>
-                  <Accordion.Header className={`mb-1 ${styles.title}`}>
-                    {`Cluster name : ${cluster.cluster}`}
-                  </Accordion.Header>
-                  <Accordion.Body className={styles.title_text}>
-                    Events
-                    <div>
-                      <div>
-                        {cluster.events
-                          && cluster.events.map((event) => (
-                            <Accordion key={event.name}>
-                              <Accordion.Header
-                                className={`mb-1 ${styles.title}`}
-                              >
-                                {`Event name : ${event.name}`}
-                              </Accordion.Header>
-                              <Accordion.Body className={styles.title_text}>
-                                <Button
-                                  variant="primary"
-                                  onClick={() => {
-                                    handleShow(event);
-                                  }}
-                                >
-                                  Update the event
-                                </Button>
-                              </Accordion.Body>
-                            </Accordion>
-                          ))}
-                      </div>
-                    </div>
-                  </Accordion.Body>
-                </Accordion>
-              ))}
-          </div>
-        )}
-      </div>
+      {!isSignedIn ? (
+        <div className="m-auto w-50">
+          <h1 className="mb-4" style={{ color: '#6AE1DD' }}>
+            Admin Login
+          </h1>
+          <Form>
+            <Form.Group className="mb-3" controlId="formLogin">
+              <Form.Label>Roll no</Form.Label>
+              <Form.Control
+                className="mb-3"
+                value={rollno}
+                type="text"
+                onChange={(e) => setRollNo(e.target.value)}
+              />
+              <Form.Label>Password</Form.Label>
+              <Form.Control
+                className="mb-3"
+                value={pass}
+                type="password"
+                onChange={(e) => setPass(e.target.value)}
+              />
+              {showIsWrongText()}
+              <Button variant="primary" onClick={handleSubmit}>
+                Login
+              </Button>
+            </Form.Group>
+          </Form>
+        </div>
+      ) : (
+        <Dashboard />
+      )}
     </Container>
   );
 }
